G00, G01 unsupported modes
I'm using gerbv just for inspection of finished PCBs in order to find minor errors before manufacturing them but, as you know, there are unsupported G modes which are affecting my PCBs when I use non-standard drills (e.g. oblong holes).

In spite of I'm not use to program in this enviroment, I made some minor modifications in order to:

get a real line counter (thru a new global variable f_line in drill and new functions to read characters from files (*))
draw oblongs ("oval" type holes) on the screen (as I just need it for a visual inspection, for now it's not important to export or print them)

(*) For now just for debug but the patch fixed the faulty line counter without modifications to the main code.